put code in right place
